refactor: use `int` instead of `short` for user-friendly refactor: add BlockContainer

+++ Allay-API/src/main/java/org/allaymc/api/blockentity/component/BlockEntityFurnaceBaseComponent.java
@@ -9,11 +9,11 @@ import org.allaymc.api.blockentity.component.common.BlockEntityBaseComponent;
  * @author daoge_cmd
  */
 public interface BlockEntityFurnaceBaseComponent extends BlockEntityBaseComponent {
-    short getBurnTime();
+    int getBurnTime();
 
-    short getCookTime();
+    int getCookTime();
 
-    short getBurnDuration();
+    int getBurnDuration();
 
     float getStoredXP();
